Refactor Next.js example: extract HomeContent into a new HomePage component, implement HomeFilters for search functionality, and update icon definitions in iconsShadcn.ts.
